unify pycuvec & helpers, minor reordering
